script/apka/requirements.txt
pygame
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
gtts